Updates related to CheckStyle recommended changes.
