+++ MavenWebDriverTest/src/test/java/dataproviders/ListDataProvider.java
@@ -1,4 +1,4 @@
-package dataProviders;
+package dataproviders;
 
 import org.testng.annotations.DataProvider;
 import java.lang.reflect.Method;
