+++ MavenWebDriverTest/src/test/java/dataproviders/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Provides classes that can be used as TestNG data providers
+ * for the test cases.
+ *
+ */
+package dataproviders;
