+++ MavenWebDriverTest/src/test/resources/data/SearchData.csv
@@ -2,4 +2,3 @@ SearchText,SearchResult,Search Result - Yahoo
 Xgames,http://xgames.espn.com/,X Games and action sports videos
 Rolex 24 Wikipedia,https://en.wikipedia.org/wiki/24_Hours_of_Daytona,24 Hours of Daytona - Wikipedia
 Tom Brokaw IMDb,http://www.imdb.com/name/nm0111232/,Tom Brokaw - IMDb
-Trust No One, set to fail, set to fail
